refactor: いいね機能をLikeButtonコンポーネントとして切り出し
